Wait action, basic time system, whitespace removed
First time using magit for emacs! Seems cool so far.
So yeah, the actual commit is pretty basic: A new action for waiting via
the '.' key, which simply passes the turn without doing anything; the
basics of the time system which will form the basis of NPC plans; the
removal of all unnecessary whitespace via the
'delete-trailing-whitespace' command in emacs. Might have to add that to
my save hook for python files!