Better town generation, beginning of plan system
Tweaked some of the variables so the generated town looks a little
better than it did before (more space around the edges of the map, fewer
random decorations).

Started to implement the NPC plan system. So far, I've created the Plan
class that each NPC will have, along with the generic 'PlanEntry' class
from which specific plan entries will inherit. The PlanEntries will be
things like 'VisitNeighbour', 'MoveToOtherRoom', 'KillFriend', and
whatnot. The current structure may not be ideal, and I may have to move
forward with moving things out of this monolithic main file before I go
truly insane.

Currently, the Plan object can be called each turn to see if there's a
PlanEntry at the current time. If there is, it will execute the plan. If
the plan can't be executed, we see if it can be rescheduled (things like
VisitNeighbour can be dropped, but things like KillFriend cannot) and
perform the reschedule if possible.

Next steps with Plans: Interface for adding PlanEntries, Plan
generation, and of course, PlanEntry subclasses.