VisitNeighbour prototype implementation
A couple of little fixes to some dodgy typos that happen when you're not
testing code as you write it (oops).

Implementation and testing of the VisitNeighbour PlanEntry. Basically,
all it does is take a target Square and pathfinds the associated NPC
through the front door. It doesn't stop him from immediately leaving,
which it probably should. I may introduce NPC states before continuing
with PlanEntries. NPC states will be things like "VisitingHouse",
"StayingIn", "Hunting", "Scared", "Dead".. You know, normal things.

This commit also highlights a bug, possibly with door placement, that
needs fixing. Not every tile is reachable according to the algorithm
we're using (because occasionally it just spins for ages and ages when
trying to find a path). So I'll also have to figure out why sometimes
the doors don't play nice, and maybe use a flood-fill to decide if we
need to ditch the Town and start again.