Getting ready for proper Plan generation
Moved player door opening to a function finally. More important, started
working towards generating Plans for NPCs in an automatic way. The
current method is really, really basic for a couple of reasons: One, we
only have one PlanEntry, so it's hard to generate anything clever. Two,
NPCs visit houses even if the owner of the house isn't going to be
there. Both of these will be sorted out in the future.