Started work on the Look function, police fix
The look function is coming along nicely and isn't nearly as bad as I
thought it would be! Once I add descriptions and stuff, the Look
function will actually be useful.

Police are now lumped in with all NPCs, but NPCs are now also split off
in to 'Villagers', which is basically the word I used when I couldn't
think of another term for people who live in a gated community. This
effectively makes the Police class pointless, and we could just make
them in to NPCs that are blue @ symbols and have the default
behaviour. Maybe next commit.