Basic Notebook added, conversation expanded
Wow, this almost passes for a real commit! Here's the rundown:
1) Notebook added. Players no longer know who every NPC is (except for
fellow police officers) and have to talk to them to find out their name
and address.
2) Dialogue is a little more 'intelligent', but still needs to be moved
out and made in to some sort of tree format. It's hard to move forward
with the actual questioning until I'm done with the NPC scheduling for the
day of the murder.

So basically, the dialogue thing is probably going to be left alone
until the town expansion is done and the day scheduling is finished.